+++ src/mongodb/documents/recipe.document.ts
@@ -1,0 +1,8 @@
+import { Document } from 'mongoose';
+import { Language } from '../../common/types';
+import { DishRecipeSections } from '../../modules/dish/dish.types';
+
+export interface RecipeDocument extends Document {
+    readonly language: Language;
+    readonly recipeSections: DishRecipeSections;
+}